requirements-dev.txt
pytest==4.4.0
pytest-django==3.4.8
flake8==2.4.1
coverage==4.0
Sphinx==1.3.1
pytest-xdist==1.13.1
pytest-cov==2.3.1
beautifulsoup4
